requirements.txt
django
djangorestframework
django-simple-login
pillow